daemon/pkg/registry: move newIndexInfo to search
It's the only user of it.
